Ability to specify version but default to latest
In dockerfile, it is recommended to specify a version of tools to be installed instead of using latest. Updating the script to that it allows an option to specify a version. Usage: cat getCli.sh | sh -s -- 1.16.0 will download specific version vs cat getCli.sh | sh that downloads latest. Tested on mac